rename SwapAmountInput and remove old version
