+++ src/swap/index.ts
@@ -1,8 +1,7 @@
 // 🌲☀️🌲
 export { getSwapQuote } from './core/getSwapQuote';
-export { Swap } from './components/Swap'
+export { Swap } from './components/Swap';
 export { SwapAmountInput } from './components/SwapAmountInput';
-export { SwapAmountInputV2 } from './components/SwapAmountInputV2';
 export { SwapButton } from './components/SwapButton';
 export type {
   BuildSwapTransactionParams,
